Fixes #1442: Unable to make call if disabled media is included
Add an API pjsua_schedule_timer2() to allow application to schedule a callback function to be executed after a specified time interval. This enables app to post a delayed job which, in this case, allows the initialization of all media transport creations to finish first before we get the media transport creations result.



git-svn-id: file:///mirror/SvnSyncs/PJ/repos/pjproject/trunk@3938 bdd19603-ad22-4b21-bbe5-5c27d6d023c3